% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{chiapas_dry_forests_df}
\alias{chiapas_dry_forests_df}
\title{Productivity in Chiapas Dry Forests}
\format{
A data frame with 96 observations and 2 variables:
\describe{
  \item{richness}{Species richness (integer)}
  \item{ndvi}{Normalized Difference Vegetation Index (NDVI), a measure of vegetation productivity (numeric)}
}
}
\source{
Data taken from the ADER package version 1.5
}
\usage{
data(chiapas_dry_forests_df)
}
\description{
This dataset, chiapas_dry_forests_df, is a data frame containing data on species richness and
vegetation productivity (measured via NDVI) in the dry forests of Chiapas, Mexico. It includes
96 observations and 2 variables. These indicators are often used in ecological and environmental
analyses to assess biodiversity and landscape productivity.
}
\details{
The dataset name has been kept as 'chiapas_dry_forests_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
MexicoDataAPI package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
