% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexicodataapi-package.R
\docType{package}
\name{MexicoDataAPI}
\alias{MexicoDataAPI}
\alias{MexicoDataAPI-package}
\title{MexicoDataAPI: Access Mexican Data via APIs and Curated Datasets}
\description{
{
This package provides functions to access data from the 'World Bank API' and the 'REST Countries API', related to Mexico's economic indicators, population statistics, literacy rates, and international geopolitical information. Additionally, the package includes curated datasets related to Mexico, including air quality monitoring stations, pollution zones, income surveys, postal abbreviations, election studies, forest productivity, and demographic data by state.
}
}
\details{
MexicoDataAPI: Access Mexican Data via APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Mexican Data via APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/mexicodataapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
