% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexicodataapi-package.R
\docType{package}
\name{MexicoDataAPI}
\alias{MexicoDataAPI}
\alias{MexicoDataAPI-package}
\title{MexicoDataAPI: Access Mexican Data via APIs and Curated Datasets}
\description{
{
This package provides functions to access data from the 'World Bank API','REST Countries API' and 'Nager.Date API', covering Mexico's economic indicators, population statistics, literacy rates, international geopolitical information and official public holidays. The package also includes curated datasets related to Mexico such as air quality monitoring stations, pollution zones, income surveys, postal abbreviations, election studies and more.
}
}
\details{
MexicoDataAPI: Access Mexican Data via APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Mexican Data via APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/mexicodataapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
